Adding package parameter IsPackageExecutedRemotely to DailyPrices package so can skip the DimPlayer tables processing if called from the Weekly package.
Updated Weekly package to execute DailyPrices package so it can process this on the day it is run (and disable certain tasks that are run in this package).

Added fixture processing to include new IsScheduled column for when fixtures are removed due to FA Cup matches, but not yet rescheduled.